Little refactoring

+++ cleanup.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-docker rm -f $(docker ps --all --format '{{.Names}}' | grep wal-g_*)
+if $(docker ps --all --format '{{.Names}}' | grep wal-g_*) > /dev/null; then
+    docker rm -f $(docker ps --all --format '{{.Names}}' | grep wal-g_*)
+fi
